+++ include/phase4/engine/board/session.h
@@ -215,7 +215,7 @@ private:
 	common::FastVector<common::PieceType> m_promotedPieces;
 	common::FastVector<ZobristHashing> m_hashes;
 	common::FastVector<ZobristHashing> m_pawnHashes;
-	common::FastVector<uint8_t> m_irreversibleMovesCounts;
+	common::FastVector<size_t> m_irreversibleMovesCounts;
 	common::FastVector<common::FieldIndex> m_wallSlides;
 };
